requirements.txt
requests==2.22.0
python_telegram_bot.egg==info
python-dotenv==0.20.0
telegram==0.0.1
